// src/player/types.ts
import {Texture} from "pixi.js";

export interface DeadAssets {
    dead: Texture;
    dead_outline: Texture;
}

export interface DuckAssets {
    duck_1: Texture;
    duck_2: Texture;
}

export interface IdleAssets {
    idle_1: Texture;
    idle_2: Texture;
}

export interface RunAssets {
    run_1: Texture;
    run_2: Texture;
}